Version 0.2.0

+++ Scripts/UIManager.gd
@@ -18,25 +18,30 @@ var sprite
 func _ready():
 	$Inspector.visible = true
 	
-	$Header/SaveFile.add_filter("*.JSON ; JSON Files")
-	$Header/SaveFile.connect("confirmed", self, "_save_at_dir")
-	$Header/OpenFile.add_filter("*.JSON ; JSON Files")
-	$Header/OpenFile.connect("file_selected", self, "_open_file")
+	$SaveFile.add_filter("*.JSON ; JSON Files")
+	$OpenFile.add_filter("*.JSON ; JSON Files")
+	$ImportFile.add_filter("*.JSON ; JSON Files")
+	
+	var err: int
+	err = $SaveFile.connect("confirmed", self, "_save_at_dir")
+	err = $OpenFile.connect("file_selected", self, "_open_file")
+	err = $ImportFile.connect("file_selected", self, "_import_file")
 	
 	var file_popup = $Header/Separator/File.get_popup()
-	file_popup.connect("id_pressed", self, "_file_option_selected")
-	file_popup.add_item("Open (CTRL+O)")
+	err = file_popup.connect("id_pressed", self, "_file_option_selected")
 	file_popup.add_item("Save (CTRL+S)")
+	file_popup.add_item("Open (CTRL+O)")
+	file_popup.add_item("Import (CTRL+I)")
 	file_popup.add_item("Open Inspector (I)")
 	
 	var box_popup = $Header/Separator/Box.get_popup()
-	box_popup.connect("id_pressed", self, "_box_option_selected")
+	err = box_popup.connect("id_pressed", self, "_box_option_selected")
 	box_popup.add_item("New (A)")
 	box_popup.add_item("Delete selected (Z)")
-	box_popup.add_item("Import from another frame (CTRL+I)")
+	box_popup.add_item("Import from another frame (SHIFT+I)")
 	
 	
-	$Header/ImportFrame.connect("confirmed", self, "_import_frame_data")
+	err = $Header/ImportFrame.connect("confirmed", self, "_import_frame_data")
 	$Header/ImportFrame.register_text_enter($Header/ImportFrame/Separator/From)
 	
 	$Inspector/Container/VerticalSeparator/TypeContainer/Type.add_item("HITBOX")
@@ -44,16 +49,17 @@ func _ready():
 	$Inspector/Container/VerticalSeparator/TypeContainer/Type.add_item("PARRY")
 	$Inspector/Container/VerticalSeparator/TypeContainer/Type.add_item("TAUNT")
 	
-	$Header/Separator/SpriteSelector.connect("item_selected", self, "_change_sprite")
+	err = $Header/Separator/SpriteSelector.connect("item_selected", self, "_change_sprite")
 	load_animated_entities()
+	
+	if err != OK:
+		push_error("Error")
+		return
 	
 	h_boxes = get_tree().get_root().get_node("Canvas/Boxes")
 	sprite = get_tree().get_root().get_node("Canvas/SpriteContainer/Sprite")
 	
-	if $Header/Separator/SpriteSelector.get_item_count() > 0:
-		pass
-		
-func _process(delta):
+func _process(_delta: float) -> void:
 	if not loaded_animations and is_instance_valid(sprite):
 		animation_player = get_parent().animation_player
 		
@@ -72,45 +78,45 @@ func _process(delta):
 		
 	# Application shortcuts
 	# Box
-	if Input.is_action_just_pressed("add_new"):
+	if Input.is_action_just_pressed("add_new") and not Utils.has_popup_open():
 		_create_box()
 	
-	if Input.is_action_just_pressed("delete_selected"):
+	if Input.is_action_just_pressed("delete_selected") and not Utils.has_popup_open():
 		_delete_selected_box()
 		
-	if Input.is_action_just_pressed("open_import_menu"):
+	if Input.is_action_just_pressed("open_import_menu") and not Utils.has_popup_open():
 		_open_import_popup()
 		
 	# Animation
-	if Input.is_action_just_pressed("stop_animation") and Utils.is_playing:
+	if Input.is_action_just_pressed("stop_animation") and Utils.is_playing and not Utils.has_popup_open():
 		Utils.is_playing = false
-		print("false")
 	
-	if Input.is_action_just_pressed("play_animation") and $Header/Separator/SpriteSelector.get_item_count() > 0:
+	if Input.is_action_just_pressed("play_animation") and $Header/Separator/SpriteSelector.get_item_count() > 0 and not Utils.has_popup_open():
 		Utils.is_playing = true
 	
+	if Input.is_action_just_pressed("open_inspector") and not Utils.has_popup_open():
+		_open_inspector()
+		
 	# File
 	if Input.is_action_just_pressed("save_file"):
-		$Header/SaveFile.visible = true
-		$Header/SaveFile.invalidate()
+		$SaveFile.popup_centered()
+		$SaveFile.invalidate()
 	
 	if Input.is_action_just_pressed("open_file"):
-		$Header/OpenFile.visible = true
-		$Header/OpenFile.invalidate()
+		$OpenFile.popup_centered()
+		$OpenFile.invalidate()
 	
-	if Input.is_action_just_pressed("open_inspector"):
-		_open_inspector()
+	if Input.is_action_just_pressed("import"):
+		$ImportFile.popup_centered()
+		$ImportFile.invalidate()
 	
 func _setup_initial_sprite():
 	var sprite_path: String = ANIMATED_SCENES_PATH + animated_entities_list[0]
 	var new_sprite = load(sprite_path).instance()
 	new_sprite.name = "Sprite"
 	new_sprite.global_position = get_parent().get_node("MiddlePoint").global_position
-	print(new_sprite.get_children())
-	print(new_sprite.get_node("AnimationPlayer").is_playing())
 	new_sprite.get_node("AnimationPlayer").stop()
 	get_node("/root/Canvas/SpriteContainer").add_child(new_sprite, true)
-#	_update_sprite_selector()
 
 func load_animated_entities():
 	var animated_entities: Array = []
@@ -119,8 +125,10 @@ func load_animated_entities():
 	if not animated_entities_dir.dir_exists(ANIMATED_SCENES_PATH):
 		return # TODO: Show a tooltip explaining the needing of a "Test" folder containing the .TSCN files
 	
-	animated_entities_dir.open(ANIMATED_SCENES_PATH)
-	animated_entities_dir.list_dir_begin()
+	var err: int = animated_entities_dir.open(ANIMATED_SCENES_PATH)
+	err = animated_entities_dir.list_dir_begin()
+	if err != OK:
+		push_error("Error")
 	
 	while true:
 		var file = animated_entities_dir.get_next()
@@ -160,21 +168,25 @@ func _update_sprite_selector():
 	for animation in new_sprite_animations:
 		$Header/Separator/AnimationSelector.add_item(animation)
 	$Header/Separator/AnimationSelector.select(0)
-	pass
 		
 func _file_option_selected(ID: int):
 	match ID:
 		0:
-			$Header/OpenFile.visible = true
-			$Header/OpenFile.invalidate()
-			if $Header/SaveFile.visible:
-				$Header/SaveFile.visible = false
+			$SaveFile.popup_centered()
+			$SaveFile.invalidate()
+			if $OpenFile.visible:
+				$OpenFile.visible = false
 		1:
-			$Header/SaveFile.visible = true
-			$Header/SaveFile.invalidate()
-			if $Header/OpenFile.visible:
-				$Header/OpenFile.visible = false
-		2: _open_inspector()
+			$OpenFile.popup_centered()
+			$OpenFile.invalidate()
+			if $SaveFile.visible:
+				$SaveFile.visible = false
+		2:
+			$ImportFile.popup_centered()
+			$ImportFile.invalidate()
+			if $SaveFile.visible:
+				$SaveFile.visible = false
+		3: _open_inspector()
 				
 func _box_option_selected(ID: int):
 	match ID:
@@ -183,20 +195,33 @@ func _box_option_selected(ID: int):
 		2: _open_import_popup()
 
 func _save_at_dir():
-	var dir: String = $Header/SaveFile.get_current_path()
+	var dir: String = $SaveFile.get_current_path()
 	var file = File.new()
 	file.open(dir + Utils.JSON_SUFIX, File.WRITE)
 	file.store_string(to_json(Utils.boxes_data))
 	file.close()
 
 func _open_file(dir: String):
+	Utils.wipe_data()
 	var file = File.new()
 	if not file.file_exists(dir):
 		return
 	
 	file.open(dir, File.READ)
 	Utils.boxes_data = parse_json(file.get_as_text())
-	_import_frame_data()
+	var frame_string: String = $Header/Separator/CurrentFrameLabel.text.replace("Frame: ", "")
+	_import_frame_data(frame_string)
+	file.close()
+
+func _import_file(dir: String) -> void:
+	var file = File.new()
+	if not file.file_exists(dir):
+		return
+	
+	file.open(dir, File.READ)
+	Utils.boxes_data = parse_json(file.get_as_text())
+	var frame_string: String = $Header/Separator/CurrentFrameLabel.text.replace("Frame: ", "")
+	_import_frame_data(frame_string)
 	file.close()
 
 func _open_inspector():
@@ -207,11 +232,13 @@ func _on_AnimationSelector_item_selected(ID):
 		if node is AnimationPlayer:
 			animation_player = node
 
-	animation_player.current_animation = animation_player.get_animation_list()[ID]
 	animation_player.seek(0, true)
+	animation_player.current_animation = animation_player.get_animation_list()[ID]
 	if animation_player.is_playing():
 		animation_player.stop()
-	Utils.is_playing = false
+		Utils.is_playing = false
+	Utils.seek_frame(0)
+	Utils.create_stored_boxes()
 
 
 func _on_PlayBtn_pressed():
@@ -234,6 +261,7 @@ func _create_box():
 	box_instance.created_manually = true
 	boxes_parent.add_child(box_instance)
 	box_instance.rect_global_position = get_tree().get_root().get_node("Canvas/MiddlePoint").global_position
+	Utils.save_data(Utils.get_animation_frame())
 
 func _delete_selected_box():
 	var boxes_parent = get_tree().get_root().get_node("Canvas/Boxes")
@@ -255,15 +283,15 @@ func _delete_selected_box():
 		Utils.clear_selected_box()
 	
 	# Save updated data
-	Utils.save_data(int(animation_player.current_animation_position * 10))
+	Utils.save_data(int(Utils.animation_pos * 10))
 
 func _open_import_popup():
 	$Header/ImportFrame.popup()
 
-func _import_frame_data() -> void:
+func _import_frame_data(frame: String = "") -> void:
 	# Save selected frame	
-	var selected_frame: String = str(int($Header/ImportFrame/Separator/From.text))
-	
+	var selected_frame: String = str(int($Header/ImportFrame/Separator/From.text)) if frame == "" else frame
+	print(selected_frame)
 	# Clears the import frame text
 	$Header/ImportFrame/Separator/From.text = ""
